requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep Learning
tensorflow>=2.15.0
keras>=2.15.0

# Time Series
statsmodels>=0.14.0

# Technical Analysis
pandas-ta>=0.3.14b
ta-lib>=0.4.28

# NLP & Sentiment
transformers>=4.35.0
torch>=2.1.0

# Data Sources
yfinance>=0.2.32
alpha-vantage>=2.3.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
mplfinance>=0.12.10

# Web Framework (for Phoenix)
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Database (optional)
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
